ath10k: Correctly report tx-rate of 48Mbps.
48Mbps is special in that it has a rate-code of 0x0, and flags == 0
means success, so the old code to check whether tx-rate status was set
or not did not properly match.  So, remember if we have received *any*
tx-rate info, and if so, we can assume that we will continue receiving it
since that is how the CT firmware works (and stock firmware never sets
any tx-rate info at all).

Signed-off-by: Ben Greear <greearb@candelatech.com>